Add original code example for mutable global data
`defaultOwner` is mutable, global data that can be mutated from anywhere
in the code.

The scope of access here might be narrow, but as the application grows
it will become harder to control access to this.

Functions are a lot easier to move around than data - you'll have to
rename all the references to the data to keep the code working.